+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Added
     - ``cleanappsettings`` management command (#374)
     - ``exclude_inherited`` argument in ``Project.get_delegates()`` (#595)
     - Optional options for app settings of type ``STRING`` and ``INTEGER`` (#592)
+    - Placeholder visible for app setting form fields fields of type ``STRING`` and ``INTEGER`` (#584)
 
 Changed
 -------